Added updater-from-file implementation, including a call to fortran code. Synched it up with both schedulers.
As it stands, the TW code runs as-fast-as possible; as a result, it
isn't entirely clear what's going on in the simulation. The cracks all
update very quickly, but initially loading the cracks takes a long time,
since the fortran code is being called. As a result, the cracks just
suddenly appear with no evidence as to how they grew.